backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector==0.3.6
openai==1.58.1
redis==5.2.1
boto3==1.35.84
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
cors==1.0.1
fastapi-cors==0.0.6
pillow==11.0.0
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1